authStateChanges runs just in main & few bugs.
Now when user logout he won't be able to go back to the HomeScreen with back button.
Solved error:
- Login requires to tap twice.

Now authStateChanges which is a listener of events from the firebase instance is being used only in the main when running the application for the first time:
- authStateChanges listen for events and when used inside a widget can cause to triggers the function.
For example if a widget is mount and has a button with this function then It doesn't matter if you don't press the button, the code inside the listener will run again. This can cause some troubles.

SignIn and Signup methods from the forms now redirect to other Screens inside their respective method.